# [GSoC] Reclaim unused KDE error tolerance