--- a/src/mlpack/tests/kde_test.cpp
+++ b/src/mlpack/tests/kde_test.cpp
@@ -135,12 +135,12 @@ BOOST_AUTO_TEST_CASE(KDETreeAsArguments)
  */
 BOOST_AUTO_TEST_CASE(GaussianKDEBruteForceTest)
 {
-  arma::mat reference = arma::randu(2, 200);
-  arma::mat query = arma::randu(2, 60);
+  arma::mat reference = arma::randu(2, 3000);
+  arma::mat query = arma::randu(2, 100);
   arma::vec bfEstimations = arma::vec(query.n_cols, arma::fill::zeros);
   arma::vec treeEstimations = arma::vec(query.n_cols, arma::fill::zeros);
-  const double kernelBandwidth = 0.3;
-  const double relError = 0.01;
+  const double kernelBandwidth = 0.12;
+  const double relError = 0.05;
 
   // Brute force KDE.
   GaussianKernel kernel(kernelBandwidth);
@@ -169,12 +169,12 @@ BOOST_AUTO_TEST_CASE(GaussianKDEBruteForceTest)
  */
 BOOST_AUTO_TEST_CASE(GaussianSingleKDEBruteForceTest)
 {
-  arma::mat reference = arma::randu(2, 300);
+  arma::mat reference = arma::randu(2, 3000);
   arma::mat query = arma::randu(2, 100);
   arma::vec bfEstimations = arma::vec(query.n_cols, arma::fill::zeros);
   arma::vec treeEstimations = arma::vec(query.n_cols, arma::fill::zeros);
   const double kernelBandwidth = 0.3;
-  const double relError = 0.01;
+  const double relError = 0.04;
 
   // Brute force KDE.
   GaussianKernel kernel(kernelBandwidth);