This PR is an implementation of #1954.