@robertohueso 2019-08-20
This PR implementation doesn't only reclaim absolute error tolerance but also relative error tolerance.

After 1000+ tests with different random seeds I only got the error fixed in #1979 (which should be solved once it's merged), so I think this is ready for review :)