@rcurtin 2019-08-21
Sorry for so many comments in one night. :)  I thought about how the bounds should relax for both relative and absolute error, considering just the case where we exhaustively computed the KDE estimate for all the points in one node (N_r1) and want to prune another node (N_r2).

Let's take the absolute-error-only case first, because that's a little bit easier.

In this case, we require that our final kernel density estimate `K(q)` is such that

```
K(q) - K'(q) <= \epsilon.
```

and, the way we enforce this is by allowing each point to have no more than `\epsilon / | S_r |` points (where `S_r` is the reference set).  So, if we exhaustively computed the kernel density estimate exactly for every descendant point of `N_r1`, then in our error budget we have not used `| N_r1 | * \epsilon / | S_r |` error.  We could use this in our next pruning decision, where instead of pruning with the rule

```
| k_max(q, N_r2) - k_min(q, N_r2) | <= 2 \epsilon / |S_r|
```

we have an extra absolute budget of `C = | N_r1 | * \epsilon / | S_r |` for the entire node, or, `C / | N_r2 |` per point.  That gives a pruning rule of

```
| k_max(q, N_r2) - k_min(q, N_r2) | <= 2 (\epsilon / |S_r| + C / | N_r2 |).
```

So, basically, with absolute error only, we can cache `C = | N_r1 | * \epsilon / | S_r |` as the unused budget from a node where we computed exhaustively, and prune with `2 (\epsilon / | S_r | + C / | N_r2 |)` on the right hand side.

Now let's consider the relative error case.  In this setting we wish to compute the kernel density estimate such that

```
K(q) - K'(q) <= \epsilon K(q).
```

If we computed all the reference points in N_r1 exhaustively, this means that the error budget that we did not use is `C = \epsilon * \sum_{i \in N_r1} k(q, r_1)` (where the `i \in N_r1` just means "all descendant points of N_r1", because plaintext is too hard to use the notation from the tree-independent dual-tree algorithms paper :)).

Our pruning rule for N_r2 would normally be this:

```
| k_max(q, N_r2) - k_min(q, N_r2) | <= 2 \epsilon k_min(q, N_r2).
```

but for each point we have an extra allowable error of `C / | N_r2 |`.  This gives a modified pruning rule of

```
| k_max(q, N_r2) - k_min(q, N_r2) | <= 2 (\epsilon k_min(q, N_r2) + C / | N_r2 |).
```

Note that this kind of looks like a "combined" absolute and relative error, because that's exactly what it is!  Our relative error is for the node at hand, but our leftover error from before is no longer relative---it can now be treated as absolute.

Finally, we can now put these two things together for the "combined" absolute and relative error case that the code works for.  Assuming that `C` is again used for any unused error budget from `N_r1`, we get the pruning rule of

```
| k_max(q, N_r2) - k_min(q, N_r2) | <= 2 (\epsilon_abs / | S_r | + C / | N_r2 | + \epsilon_rel k_min(q, N_r2)
```

So, I think the general strategy of what you wrote is just the same, but I think that we can just adjust slightly how we are computing `accumError`, then when we use it, we just need to divide it by `referenceNode.NumDescendants()`.

As always please double-check my math to see if anything is wrong with it.  It's possible I overlooked something or made an error. :)  And let me know if anything was unclear.  I was a little bit lazy with the notation in places, but hope it still makes sense.