@rcurtin 2019-08-29
I did a little extra benchmarking for absolute error.  I used again the covertype dataset with absolute error of 0.1 and a bandwidth for the Gaussian kernel of 500.

```
            time      base cases      scored nodes

old         15682.0s  214.79B         19.74B
bound       15429.9s  210.01B         19.38B
new         11468.0s  154.96B         16.88B
with-mc'    12991.6s  154.30B         16.80B
```

I wrote `with-mc'` not `with-mc` because I am trying to tune the parameters to see if I can get any speedup.  For that trial I tried `--mc_entry_coef 8 --initial_sample_size 25 --mc_break_coef 0.15`.  I'll try some more as time goes on and see what happens. :)