@rcurtin 2019-08-29
I'm really happy to have this merged.  If you want to keep benchmarking, I think maybe we can figure out where this code is fast (and where it's slow), and maybe see the way to further improvements, if you like. :+1: