@rcurtin 2019-09-07
@robertohueso I did just a little bit more benchmarking using Monte Carlo sampling and was able to get some improvements.  As I swept `initial_sample_size`, I saw significant changes in runtime:

```
iss = 5      6928.089s
iss = 15    6222.226s
iss = 25   13367.768s
```

so by tuning the parameters, it seems like better performance can be found.  But that probably differs for every dataset.